Move code for "Add Node" to separate file
